diff: use intmax_t for column widths
Don’t represent column widths and tab stops as size_t, since they
are not object sizes.  Instead, use intmax_t.  This is better than
uintmax_t since gcc -fsanitize=undefined can catch overflows.
* src/diff.c (main):
* src/diff.h (tabsize, sdiff_half_width, sdiff_column2_offset):
* src/io.c (find_and_hash_each_line):
* src/side.c (tab_from_to, print_half_line, print_1sdiff_line):
* src/util.c (lines_differ, output_1_line):
Use intmax_t for column numbers, not size_t.
* src/io.c (find_and_hash_each_line): Count only columns since the
last tab stop, to avoid signed integer overflow.  (The old code
had this possibility on odd but valid platforms where SIZE_MAX <=
INT_MAX.)
* src/util.c (lines_differ, output_1_line): Use a separate counter
for tab stops, so that the column count counts only columns since
the last tab stop.  This avoids integer overflow when tabsize is
large.  (The old code mishandled this situation.)

+++ src/diff.h
@@ -186,7 +186,7 @@ XTERN bool brief;
 XTERN bool expand_tabs;
 
 /* Number of columns between tab stops.  */
-XTERN size_t tabsize;
+XTERN intmax_t tabsize;
 
 /* Use a tab in the output, rather than a space, before the text of an
    input line, so as to keep the proper alignment in the input line
@@ -223,8 +223,8 @@ XTERN bool left_column;
 XTERN bool suppress_common_lines;
 
 /* The half line width and column 2 offset for OUTPUT_SDIFF.  */
-XTERN size_t sdiff_half_width;
-XTERN size_t sdiff_column2_offset;
+XTERN intmax_t sdiff_half_width;
+XTERN intmax_t sdiff_column2_offset;
 
 /* String containing all the command options diff received,
    with spaces between and at the beginning but none at the end.
